Move Fish prompt config to general location
The fish prompt had previously been removed with a universal variable,
but that is annoying to sync between devices. So, at one point I had
configured it in Nix for Titanium, but then I lost that config on my
other devices. To keep it even on non-Nix systems, this moves the
configuration change to config.nix.